Plugin overhaul, general cleanup
