+++ config/locales/es.yml
@@ -1,7 +1,7 @@
 # Spanish strings go here for Rails i18n
 es:
   label_setting_examples: "Vea algunos ejemplos de configuración en "
-  label_example_url: http://gatatac.net/projects/redminewikistartpage/wiki
+  label_example_url: https://github.com/txinto/redmine_startpage
   label_startpage_active: Activado?
   label_startpage_controller: Controlador
   label_startpage_action: Acción
