+++ init.rb
@@ -1,12 +1,13 @@
 require 'redmine'
 
+require_dependency 'welcome_controller_patch'
+
 Redmine::Plugin.register :redmine_startpage do
   name 'Redmine Startpage plugin'
   author 'Txinto Vaz'
   description 'This is a plugin for Redmine.  It allows the user to select almost any redmine sub page as start page for a Redmine website'
-  version '0.0.2'
-  url 'http://gatatac.net/projects/redminewikistartpage'
-  author_url 'http://gatatac.net/users/3'
+  version '0.1.0'
+  url 'https://github.com/txinto/redmine_startpage'
   
   settings :default => {
     'startpage_active' => :false,
