--- lang/en.yml
@@ -1,2 +1,0 @@
-# English strings go here
-my_label: "My label"
